+++ app/src/main/java/woowacourse/shopping/ui/cart/listener/OnAllCartItemSelectedListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.cart
+package woowacourse.shopping.ui.cart.listener
 
 interface OnAllCartItemSelectedListener {
     fun selectedAll()
